Refactored structure, shared utility functions, components and hooks
