+++ react/src/lib/shared-types/point.ts
